Code clean-up: Remove ancient Dialog class
Use Java's assert instead of our own assertion

+++ edu.rice.cs.hpcdata/src/edu/rice/cs/hpcdata/db/IdTuple.java
@@ -260,7 +260,7 @@ public class IdTuple
 				if (i==1) {
 					str += ".";
 				} else if (i>1) {
-					lblIndex = (long) (Math.pow(10, level-i) * logicalIndexes[i]);
+					lblIndex = (long) (Math.pow(10, (double)level-i) * logicalIndexes[i]);
 				}
 				str += lblIndex;
 			}
